[CHORE] 뷰 컴포넌트화 및 파일 분리
