+++ src/assets/css/ootd/ootd-detail.css
