--- src/assets/ootd.css
@@ -1,216 +1,0 @@
-.ootd-container {
-    display: inline-flex;
-    flex-direction: column;
-    align-items: center;
-    gap: 1.1875rem;
-
-    overflow-y: hidden;
-}
-
-.ootd-header-container {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    gap: 2.8125rem;
-}
-
-.ootd-header-title-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    gap: 0.625rem;
-}
-
-.ootd-header-title-text {
-    color: var(--Dark-Red, #C22727);
-    font-family: The Jamsil;
-    font-size: 3.125rem;
-    font-style: normal;
-    font-weight: 700;
-    line-height: normal;
-}
-
-.ootd-header-bar-wrapper {
-    display: flex;
-    align-items: center;
-    gap: 64.75rem;
-}
-
-.ootd-header-sort-wrapper {
-    display: flex;
-    align-items: flex-start;
-}
-
-.ootd-sort-wrapper {
-    display: flex;
-    padding: 0.625rem;
-    justify-content: center;
-    align-items: center;
-}
-
-.ootd-sort-text {
-    color: var(--Grayscale7, #000);
-    text-align: right;
-    font-family: The Jamsil;
-    font-size: 1.5625rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-}
-
-.ootd-sort-text:hover {
-    cursor: pointer;
-}
-
-.ootd-sort-text.selected {
-    color: var(--Dark-Red, #C22727);
-}
-
-.ootd-sort-line-wrapper {
-    display: flex;
-    padding: 0.625rem;
-    justify-content: center;
-    align-items: center;
-}
-
-.ootd-sort-line {
-    color: var(--Grayscale7, #000);
-    text-align: right;
-    font-family: The Jamsil;
-    font-size: 1.5625rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-}
-
-.ootd-header-write-btn-wrapper {
-    display: flex;
-    height: 1.5rem;
-    padding: 0.625rem;
-    justify-content: center;
-    align-items: center;
-
-    background: #000;
-}
-
-.ootd-header-write-btn {
-    color: var(--Grayscale1, #FFF);
-    font-family: The Jamsil;
-    font-size: 1.25rem;
-    font-style: normal;
-    font-weight: 300;
-    line-height: normal;
-}
-
-.ootd-header-write-btn-wrapper:hover {
-    cursor: pointer;
-}
-
-.ootd-body-container {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    align-items: flex-start;
-    justify-content: center;
-    gap: 1.1875rem;
-}
-
-.ootd-post-card-line-wrapper {
-    display: flex;
-    align-items: flex-start;
-    gap: 1.375rem;
-}
-
-.ootd-post-card-wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-end;
-    gap: 0.1875rem;
-}
-
-.ootd-post-card-image-wrapper:hover {
-    cursor: pointer;
-}
-
-.ootd-post-card-image-wrapper {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-}
-
-.ootd-post-card-image {
-    width: 21.375rem;
-    height: 26.9375rem;
-
-    border-radius: 0.625rem;
-}
-
-.ootd-post-card-like-wrapper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    gap: 0.625rem;
-
-    position: absolute;
-    right: 1rem;
-    bottom: 0.83125rem;
-}
-
-.ootd-post-card-like-icon {
-
-}
-
-.ootd-post-card-like-icon.selected {
-    fill: #F00;
-    fill-opacity: 100%;
-    stroke: none;
-}
-
-.ootd-post-card-like {
-    width: 2.5rem;
-    height: 2.29375rem;
-
-    fill: rgba(218, 218, 218, 0.40);
-    stroke-width: 2px;
-    stroke: #FFF;
-}
-
-.ootd-post-card-view-wrapper {
-    display: flex;
-    align-items: flex-start;
-    gap: 0.25rem;
-}
-
-.ootd-post-card-like-view-wrapper {
-    display: flex;
-    align-items: flex-start;
-    gap: 0.25rem;
-}
-
-.ootd-post-card-like-view-count-text {
-    color: var(--Red, #F00);
-    text-align: center;
-    font-family: The Jamsil;
-    font-size: 0.9375rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-}
-
-.ootd-post-card-count-view-wrapper {
-    display: flex;
-    align-items: center;
-    gap: 0.25rem;
-}
-
-.ootd-post-card-count-view-count-text {
-    color: var(--Grayscale4, #C6C6C6);
-    text-align: center;
-    font-family: The Jamsil;
-    font-size: 0.9375rem;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-}
